backend/requirements.txt
Flask==3.1.1
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
mongoengine==0.28.2
pymongo==4.6.1
python-dotenv==1.0.0
bcrypt==4.1.2
marshmallow==3.20.2
flask_pymongo==3.0.1
numpy
scikit-image
matplotlib
opencv-python
tqdm
h5py
torch
torchvision
